Move watchdog count to global data store
The watchdog functions are called during board_init_f, where the
.bss section are not initialized yet. The functions refers to a
watchdog count static variable, which are stored in the .bss section.
As the .bss section and .rel.sym (relocation table) section share
the same memory space, writing to the watchdog count variable corrupts
the relocation table, causing the relocation function to break with
data abort as it tries to access wrong memory.

To work around this, we moved the watchdog count variable to global
data store. This is not a good practice, as we want to keep global
data store as small as possible. Other boards uses RTC registers to
store their watchdog related variables, however on iOBC we doesn't
have RTC.

iOBC documentation states that to properly kick the watchdog we can't
do it too often, but they don't specify it